backend/pyproject.toml
[tool.poetry]
name = "calhacks-backend"
version = "0.1.0"
description = ""
authors = ["Oliver Ni <oliver.ni@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
starlette = "^0.21.0"
Pillow = "^9.2.0"
uvicorn = "^0.18.3"
python-multipart = "^0.0.5"
transformers = {version = "^4.23.1", extras = ["torch"]}

[tool.poetry.dev-dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
